requirements.txt
requests
bs4
pandas
pymongo
dnspython
python-dotenv
urllib